add: support mobile view
