+++ src/lang/en.json
@@ -130,6 +130,7 @@
     "landing_menu_6": "News",
     "landing_menu_7": "Products",
     "landing_menu_8": "Pricing",
+    "landing_menu_9": "Home",
     "button_1": "Login",
     "button_2": "Logout",
     "button_3": "Dashboard",
