+++ src/lang/id.json
@@ -131,6 +131,7 @@
     "landing_menu_6": "Berita",
     "landing_menu_7": "Produk",
     "landing_menu_8": "Pricing",
+    "landing_menu_9": "Home",
     "button_1": "Masuk",
     "button_2": "Keluar",
     "button_3": "Dashboard",
